Refactring SmartPointers WIP

+++ src/mods/mod_level_editor.cpp
@@ -1044,7 +1044,13 @@ void ModLevelEditor::handleLevelInputState(bool does_level_exist, int level_code
  */
 void ModLevelEditor::createNewLevel(int level_code)
 {
-    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(m_session_data->m_user_database);
+    session_ptr session = getLockedSession();
+    if (!session) 
+    {
+        return;
+    }
+    
+    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(getUserDatabase());
     access_level_ptr new_level = std::make_shared<AccessLevel>();
 
     // Check if Level Already Exists.
@@ -1071,7 +1077,13 @@ void ModLevelEditor::createNewLevel(int level_code)
  */
 void ModLevelEditor::deleteExistingLevel(int level_code)
 {
-    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(m_session_data->m_user_database);
+    session_ptr session = getLockedSession();
+    if (!session)
+    {
+        return;
+    }
+    
+    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(getUserDatabase());
     access_level_ptr existing_level = level_dao->getAccessLevelByLevel(level_code);
 
     if(existing_level->iId == -1 || !level_dao->deleteRecord(existing_level->iId))
@@ -1088,7 +1100,13 @@ void ModLevelEditor::deleteExistingLevel(int level_code)
  */
 void ModLevelEditor::copyExistingLevel(int level_code)
 {
-    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(m_session_data->m_user_database);
+    session_ptr session = getLockedSession();
+    if (!session)
+    {
+        return;
+    }
+    
+    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(getUserDatabase());
     access_level_ptr existing_level = level_dao->getAccessLevelByLevel(m_current_level);
 
     if(checkLevelExistsByLevel(level_code))
@@ -1113,7 +1131,7 @@ void ModLevelEditor::copyExistingLevel(int level_code)
  */
 void ModLevelEditor::saveLevelChanges()
 {
-    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(m_session_data->m_user_database);
+    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(getUserDatabase());
     access_level_ptr existing_level = nullptr;
 
     for(unsigned int i = 0; i < m_loaded_levels.size(); i++)
@@ -1169,7 +1187,7 @@ access_level_ptr ModLevelEditor::getCurrentLevel()
  */
 std::string ModLevelEditor::displayLevelList()
 {
-    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(m_session_data->m_user_database);
+    access_level_dao_ptr level_dao = std::make_shared<AccessLevelDao>(getUserDatabase());
 
     // Clear All Levels
     if(m_loaded_levels.size() > 0)
